Extract rewardable logic into new util file
