+++ src/stores/utils/contract.js
@@ -1,14 +1,8 @@
 import BN from 'bignumber.js';
 import { fromDecimals } from './decimals'
 import { fromWei } from 'web3-utils'
-import { FEE_MANAGER_MODE } from './bridgeMode'
 
 export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
-
-export const validFee = (fee) => {
-  const zeroBN = new BN(0)
-  return !zeroBN.eq(fee)
-}
 
 export const getMaxPerTxLimit = async (contract,decimals) => {
   const maxPerTx = await contract.methods.maxPerTx().call()
@@ -98,80 +92,6 @@ export const getHomeFee = async (contract) => {
 export const getForeignFee = async (contract) => {
   const feeInWei = await contract.methods.getForeignFee().call()
   return new BN(fromWei(feeInWei.toString()))
-}
-
-export const getFeeToApply = (homeFeeManager, foreignFeeManager, homeToForeignDirection) => {
-  const data = getRewardableData(homeFeeManager, foreignFeeManager)
-  return homeToForeignDirection ? data.homeFee : data.foreignFee
-}
-
-export const getRewardableData = (homeFeeManager, foreignFeeManager) => {
-  if(homeFeeManager.feeManagerMode === FEE_MANAGER_MODE.BOTH_DIRECTIONS) {
-    return {
-      homeFee: homeFeeManager.homeFee,
-      foreignFee: homeFeeManager.foreignFee,
-      homeHistoricFee: homeFeeManager.homeHistoricFee,
-      foreignHistoricFee: homeFeeManager.foreignHistoricFee,
-      depositSymbol: 'home',
-      withdrawSymbol: 'home',
-      displayDeposit: true,
-      displayWithdraw: true
-    }
-  } else if(homeFeeManager.feeManagerMode === FEE_MANAGER_MODE.ONE_DIRECTION
-    && foreignFeeManager.feeManagerMode === FEE_MANAGER_MODE.ONE_DIRECTION) {
-    return {
-      homeFee: foreignFeeManager.homeFee,
-      foreignFee: homeFeeManager.foreignFee,
-      homeHistoricFee: foreignFeeManager.homeHistoricFee,
-      foreignHistoricFee: homeFeeManager.foreignHistoricFee,
-      depositSymbol: 'foreign',
-      withdrawSymbol: 'home',
-      displayDeposit: true,
-      displayWithdraw: true
-    }
-  } else if(homeFeeManager.feeManagerMode === FEE_MANAGER_MODE.ONE_DIRECTION
-    && foreignFeeManager.feeManagerMode === FEE_MANAGER_MODE.UNDEFINED) {
-    return {
-      homeFee: new BN(0),
-      foreignFee: homeFeeManager.foreignFee,
-      homeHistoricFee: [],
-      foreignHistoricFee: homeFeeManager.foreignHistoricFee,
-      depositSymbol: '',
-      withdrawSymbol: 'home',
-      displayDeposit: false,
-      displayWithdraw: true
-    }
-  } else if(homeFeeManager.feeManagerMode === FEE_MANAGER_MODE.UNDEFINED
-    && foreignFeeManager.feeManagerMode === FEE_MANAGER_MODE.ONE_DIRECTION) {
-    return {
-      homeFee: foreignFeeManager.homeFee,
-      foreignFee: new BN(0),
-      homeHistoricFee: foreignFeeManager.homeHistoricFee,
-      foreignHistoricFee: [],
-      depositSymbol: 'foreign',
-      withdrawSymbol: '',
-      displayDeposit: true,
-      displayWithdraw: false
-    }
-  } else {
-    return {
-      homeFee: new BN(0),
-      foreignFee: new BN(0),
-      depositSymbol: '',
-      withdrawSymbol: '',
-      displayDeposit: false,
-      displayWithdraw: false
-    }
-  }
-}
-
-export const getFeeAtBlock = (feeArray, blockNumber) => {
-  for (let i = feeArray.length - 1; i >= 0; i--) {
-    if (blockNumber > feeArray[i].blockNumber) {
-      return feeArray[i].fee
-    }
-  }
-  return new BN(0)
 }
 
 export const getDeployedAtBlock = async (contract) => {
